requirements.txt
Django==5.1.1
djangorestframework==3.15.2
PyMySQL==1.1.1
Pillow==10.4.0
cryptography==43.0.1
djangorestframework-simplejwt==5.3.1
datetime==5.5
django-cors-headers==4.5.0
channels==4.1.0
channels-redis==4.2.0
django-channels-jwt==0.0.3
daphne==4.1.2